refactor(config,internal/yamlmerger): Migrate YAML merging functions
This commit moves the YAML merging functions from the feeder_yaml
file to a separate internal package from where they are exported.
