+++ config/feeder_yaml.go
@@ -38,6 +38,8 @@ import (
 	"path/filepath"
 
 	"gopkg.in/yaml.v3"
+
+	"kraftkit.sh/internal/yamlmerger"
 )
 
 // YamlFeeder feeds using a YAML file.
@@ -111,7 +113,7 @@ func (yf YamlFeeder) Write(structure interface{}, merge bool) error {
 
 	// When kind is 0, it is an uninitialized YAML structure (aka empty file)
 	if from.Kind != 0 && merge {
-		if err := recursiveMerge(&from, &into); err != nil {
+		if err := yamlmerger.RecursiveMerge(&from, &into); err != nil {
 			return fmt.Errorf("could not update config: %v", err)
 		}
 	}
@@ -125,62 +127,4 @@ func (yf YamlFeeder) Write(structure interface{}, merge bool) error {
 	}
 
 	return yaml.NewEncoder(f).Encode(&into)
-}
-
-// https://stackoverflow.com/a/65784135
-func recursiveMerge(from, into *yaml.Node) error {
-	if from.Kind != into.Kind {
-		return fmt.Errorf("cannot merge nodes of different kinds")
-	}
-
-	switch from.Kind {
-	case yaml.MappingNode:
-		for i := 0; i < len(from.Content); i += 2 {
-			found := false
-			for j := 0; j < len(into.Content); j += 2 {
-				if nodesEqual(from.Content[i], into.Content[j]) {
-					found = true
-					if err := recursiveMerge(from.Content[i+1], into.Content[j+1]); err != nil {
-						return fmt.Errorf("at key " + from.Content[i].Value + ": " + err.Error())
-					}
-					break
-				}
-			}
-			if !found {
-				into.Content = append(into.Content, from.Content[i:i+2]...)
-			}
-		}
-	case yaml.ScalarNode:
-		// SA4006 these variables represent pointers and are propagated outside of `recursiveMerge`
-		into = from //nolint:staticcheck
-	case yaml.SequenceNode:
-		for _, fromItem := range from.Content {
-			foundFrom := false
-			for _, intoItem := range into.Content {
-				if fromItem.Value == intoItem.Value {
-					foundFrom = true
-				}
-			}
-			if !foundFrom {
-				into.Content = append(into.Content, fromItem)
-			}
-		}
-	case yaml.DocumentNode:
-		if err := recursiveMerge(from.Content[0], into.Content[0]); err != nil {
-			return err
-		}
-	default:
-		return fmt.Errorf("can only merge mapping, sequence and scalar nodes")
-	}
-
-	return nil
-}
-
-func nodesEqual(l, r *yaml.Node) bool {
-	if l.Kind == yaml.ScalarNode && r.Kind == yaml.ScalarNode {
-		return l.Value == r.Value
-	}
-
-	// panic("equals on non-scalars not implemented!")
-	return false
 }
